Add static TranslationService.t() instead of t() on FObject
