+++ src/foam/nanos/approval/ApprovalRequestClassification.js
@@ -12,6 +12,10 @@
     'translationService'
   ],
 
+  javaImports: [
+    'static foam.i18n.TranslationService.t'
+  ],
+
   properties: [
     {
       class: 'String',
